# internal/api/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Motus API
  version: "0.1.0"
  description: GPS tracking and fleet management API
servers:
  - url: /
    description: This server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session cookie set after login
    bearerAuth:
      type: http
      scheme: bearer
      description: API key token
    xAuthToken:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: iOS PWA fallback — pass session_id value here
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "resource not found"
...
    Event:
      type: object
      required: [id, deviceId, type, eventTime]
      properties:
        id: { type: integer, format: int64 }
        deviceId: { type: integer, format: int64 }
        geofenceId: { type: integer, format: int64, nullable: true }
        type: { type: string }
        positionId: { type: integer, format: int64, nullable: true }
        eventTime: { type: string, format: date-time }
        attributes:
          type: object
          additionalProperties: {}
...
paths:
...
  /api/events:
    get:
      operationId: listEvents
      summary: List events for the authenticated user's devices
      tags: [events]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      parameters:
        - name: deviceId
          in: query
          schema: { type: integer, format: int64 }
        - name: type
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /api/reports/events:
    get:
      operationId: reportEvents
      summary: Report events in a time range (same as listEvents, Traccar-compatible path)
      tags: [reports]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      parameters:
        - name: deviceId
          in: query
          schema: { type: integer, format: int64 }
        - name: type
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Event report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        "401":
          $ref: '#/components/responses/Unauthorized'
...